Renamed some files to match convention.
