+++ server/spec/lifecycle/death.spec.js
