+++ server/src/lifecycle/hp-init.js
